Quick clean up
