+++ index.html
@@ -4,15 +4,9 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <!--New description for better SEO-->
     <meta name="description"
         content="Looking for a fun way to test your knowledge? Try our online quiz with unique general knowledge questions. 
         This interactive quiz is a fun way to challenge your friends and learn new things. Play now!">
-
-        <!-- Ska ej använda "keywords" - måste läggas in i en brödtext eller i rubrikerna -->
-        <!-- Lägg till open graph meta tags. -->
-        <!--     <meta name="keywords"
-        content="online quiz, brain games, knowledge competition, learn while playing, interactive quiz."> -->
     
     <meta name="author" content="Quiz Project Team - ChasAcademy">
 
@@ -20,7 +14,7 @@
     <meta property="og:description" content="Looking for a fun way to test your knowledge? Try our online quiz with unique general knowledge questions. 
     This interactive quiz is a fun way to challenge your friends and learn new things. Play now!">
     <meta property="og:type" content="website">
-    <meta property="og:url" content="https://u03-quiz-app-grupp7.netlify.app"> <!-- Lägg till när vi deployat -->
+    <meta property="og:url" content="https://u03-quiz-app-grupp7.netlify.app">
     <meta property="og:image" content="https://ibb.co/kS8NSSZ">
 
     <link rel="stylesheet" href="./assets/css/style.css">
@@ -28,10 +22,10 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=DM+Sans:ital,opsz,wght@0,9..40,100..1000;1,9..40,100..1000&family=Poppins:ital,wght@0,100;0,200;0,300;0,400;0,500;0,600;0,700;0,800;0,900;1,100;1,200;1,300;1,400;1,500;1,600;1,700;1,800;1,900&display=swap"
           rel="stylesheet">
-    <script src="./assets/js/script1.js" defer type="module"></script>
+    <script src="./assets/js/script.js" defer type="module"></script>
     <script src="./assets/js/localStorage.js" defer type="module"></script>
 
-    <!--New title for better SEO-->
+   
     <title>Quiz Fun for Everyone - Interactive and Challenging Questions!</title>
 </head>
 
@@ -44,7 +38,7 @@
    
     <section id="start-window">
         <h2>Welcome</h2>
-         <!--New Paragraph for more description/better SEO-->
+        
         <p class="welcome-msg">Challenge your brain and learn new things</p>
         <p id="instructions">Game instructions</p>
         <fieldset>
@@ -67,7 +61,7 @@
 
 
     <section id="game-over" style="display: none;">
-        <h2>Game Over</h2>
+        <h2 id="gameOver">Game Over</h2>
         <p id="final-score">Your final score is: </p>
 
         <h2 id="finalScore">LeaderBoard</h2>
@@ -78,13 +72,6 @@
 
         <button id="restart-btn">Play Again</button>
 
-        <!--<section id="leaderboard">
-            <h2>Leaderboard</h2>
-            <p>Wow! Look at this!</p>
-            <ul id="ul-leaderboard">
-    
-            </ul>
-        </section>-->
     </section>
 
 </body>
